+++ src/plone/app/imagecropping/browser/crop.py
@@ -6,8 +6,14 @@ from plone.app.imagecropping import PAI_STORAGE_KEY
 from plone.app.imagecropping.interfaces import IImageCroppingUtils
 from plone.scale.storage import AnnotationStorage
 from zope.annotation.interfaces import IAnnotations
+from zope.event import notify
 
 import PIL.Image
+
+try:
+    from z3c.caching.purge import Purge
+except ImportError:
+    Purge = None
 
 
 class CroppingView(BrowserView):
@@ -41,6 +47,10 @@ class CroppingView(BrowserView):
 
         # store crop information in annotations
         self._store(fieldname, scale, box)
+
+        # Purge caches if needed
+        if Purge is not None:
+            notify(Purge(self.context))
 
     @property
     def _storage(self):